Moved ADX to numeric package
