+++ ta4j-core/src/main/java/org/ta4j/core/indicators/numeric/NumericIndicator.java
@@ -26,7 +26,6 @@ package org.ta4j.core.indicators.numeric;
 import org.ta4j.core.BarSeries;
 import org.ta4j.core.indicators.Indicator;
 import org.ta4j.core.indicators.RSIIndicator;
-import org.ta4j.core.indicators.adx.ADXIndicator;
 import org.ta4j.core.indicators.candles.price.ClosePriceIndicator;
 import org.ta4j.core.indicators.candles.price.HighPriceIndicator;
 import org.ta4j.core.indicators.candles.price.LowPriceIndicator;
@@ -42,6 +41,7 @@ import org.ta4j.core.indicators.helpers.MedianPriceIndicator;
 import org.ta4j.core.indicators.helpers.RunningTotalIndicator;
 import org.ta4j.core.indicators.helpers.VolumeIndicator;
 import org.ta4j.core.indicators.helpers.previous.PreviousNumericValueIndicator;
+import org.ta4j.core.indicators.numeric.adx.ADXIndicator;
 import org.ta4j.core.indicators.numeric.average.EMAIndicator;
 import org.ta4j.core.indicators.numeric.average.LWMAIndicator;
 import org.ta4j.core.indicators.numeric.average.MMAIndicator;
