+++ ta4j-core/src/main/java/org/ta4j/core/indicators/numeric/adx/PlusDIIndicator.java
@@ -21,60 +21,57 @@
  * IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN
  * CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
  */
-package org.ta4j.core.indicators.adx;
+package org.ta4j.core.indicators.numeric.adx;
 
 import java.time.Instant;
 
-import org.ta4j.core.Bar;
 import org.ta4j.core.BarSeries;
-import org.ta4j.core.indicators.SeriesRelatedNumericIndicator;
+import org.ta4j.core.indicators.numeric.ATRIndicator;
+import org.ta4j.core.indicators.numeric.NumericIndicator;
+import org.ta4j.core.indicators.numeric.average.MMAIndicator;
 import org.ta4j.core.num.Num;
 
 /**
- * +DM indicator.
+ * +DI indicator.
  *
  * <p>
  * Part of the Directional Movement System.
+ *
+ * @see <a href=
+ *     "http://stockcharts.com/school/doku.php?id=chart_school:technical_indicators:average_directional_index_adx">
+ *     http://stockcharts.com/school/doku.php?id=chart_school:technical_indicators:average_directional_index_adx</a>
+ * @see <a href=
+ *     "https://www.investopedia.com/terms/a/adx.asp">https://www.investopedia.com/terms/a/adx.asp</a>
  */
-public class PlusDMIndicator extends SeriesRelatedNumericIndicator {
+public class PlusDIIndicator extends NumericIndicator {
 
-  private Bar previousBar;
+  private final int barCount;
+  private final ATRIndicator atrIndicator;
+  private final MMAIndicator avgPlusDMIndicator;
   private Num value;
+  private int barsPassed;
   private Instant currentTick = Instant.EPOCH;
-  private boolean stable;
 
 
   /**
    * Constructor.
    *
    * @param series the bar series
+   * @param barCount the bar count for {@link #atrIndicator} and
+   *     {@link #avgPlusDMIndicator}
    */
-  public PlusDMIndicator(final BarSeries series) {
-    super(series);
+  public PlusDIIndicator(final BarSeries series, final int barCount) {
+    super(series.numFactory());
+    this.barCount = barCount;
+    this.atrIndicator = new ATRIndicator(series, barCount);
+    this.avgPlusDMIndicator = new MMAIndicator(new PlusDMIndicator(series), barCount);
   }
 
 
   protected Num calculate() {
-    final var numFactory = getBarSeries().numFactory();
-
-    if (this.previousBar == null) {
-      this.previousBar = getBarSeries().getBar();
-      return numFactory.zero();
-    }
-
-    this.stable = true;
-    final Bar prevBar = this.previousBar;
-    final Bar currentBar = getBarSeries().getBar();
-
-    final Num upMove = currentBar.highPrice().minus(prevBar.highPrice());
-    final Num downMove = prevBar.lowPrice().minus(currentBar.lowPrice());
-
-    this.previousBar = currentBar;
-    if (upMove.isGreaterThan(downMove) && upMove.isGreaterThan(numFactory.zero())) {
-      return upMove;
-    }
-
-    return numFactory.zero();
+    return this.avgPlusDMIndicator.getValue()
+        .dividedBy(this.atrIndicator.getValue())
+        .multipliedBy(getNumFactory().hundred());
   }
 
 
@@ -87,19 +84,28 @@ public class PlusDMIndicator extends SeriesRelatedNumericIndicator {
   @Override
   public void refresh(final Instant tick) {
     if (tick.isAfter(this.currentTick)) {
+      ++this.barsPassed;
+      this.atrIndicator.refresh(tick);
+      this.avgPlusDMIndicator.refresh(tick);
       this.value = calculate();
       this.currentTick = tick;
-    } else if (tick.isBefore(this.currentTick)) {
-      this.previousBar = null;
-      this.stable = false;
+    } else if (tick.isBefore(tick)) {
+      this.barsPassed = 1;
+      this.atrIndicator.refresh(tick);
+      this.avgPlusDMIndicator.refresh(tick);
       this.value = calculate();
       this.currentTick = tick;
     }
   }
 
 
-  @Override
   public boolean isStable() {
-    return this.stable;
+    return this.barsPassed >= this.barCount && this.atrIndicator.isStable() && this.avgPlusDMIndicator.isStable();
+  }
+
+
+  @Override
+  public String toString() {
+    return getClass().getSimpleName() + " " + this.atrIndicator + " " + this.avgPlusDMIndicator;
   }
 }
