+++ ta4j-core/src/test/java/org/ta4j/core/indicators/numeric/adx/MinusDMIndicatorTest.java
@@ -1,4 +1,4 @@
-/**
+/*
  * The MIT License (MIT)
  *
  * Copyright (c) 2017-2023 Ta4j Organization & respective
@@ -21,7 +21,7 @@
  * IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN
  * CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
  */
-package org.ta4j.core.indicators.adx;
+package org.ta4j.core.indicators.numeric.adx;
 
 import static org.ta4j.core.TestUtils.assertNumEquals;
 
@@ -34,9 +34,9 @@ import org.ta4j.core.mocks.MockBarSeriesBuilder;
 import org.ta4j.core.num.Num;
 import org.ta4j.core.num.NumFactory;
 
-public class PlusDMIndicatorTest extends AbstractIndicatorTest<Num> {
+public class MinusDMIndicatorTest extends AbstractIndicatorTest<Num> {
 
-  public PlusDMIndicatorTest(final NumFactory numFactory) {
+  public MinusDMIndicatorTest(final NumFactory numFactory) {
     super(numFactory);
   }
 
@@ -44,25 +44,10 @@ public class PlusDMIndicatorTest extends AbstractIndicatorTest<Num> {
   @Test
   public void zeroDirectionalMovement() {
     final var series = new MockBarSeriesBuilder().withNumFactory(this.numFactory).build();
-    final var yesterdayBar = series.barBuilder().openPrice(0).closePrice(0).highPrice(10).lowPrice(2).build();
-    final var todayBar = series.barBuilder().openPrice(0).closePrice(0).highPrice(6).lowPrice(6).build();
-    final var dup = prepareIndicator(series, yesterdayBar, todayBar);
-
-    assertNumEquals(0, dup.getValue());
-  }
-
-
-  private static PlusDMIndicator prepareIndicator(
-      final BacktestBarSeries series, final BacktestBar yesterdayBar,
-      final BacktestBar todayBar
-  ) {
-    final var dup = new PlusDMIndicator(series);
-    series.replaceStrategy(new MockStrategy(dup));
-    series.addBar(yesterdayBar);
-    series.addBar(todayBar);
-    series.advance();
-    series.advance();
-    return dup;
+    final var yesterdayBar = series.barBuilder().openPrice(0).closePrice(0).highPrice(0).lowPrice(0).build();
+    final var todayBar = series.barBuilder().openPrice(0).closePrice(0).highPrice(0).lowPrice(0).build();
+    final var down = prepareIndicator(series, yesterdayBar, todayBar);
+    assertNumEquals(0, down.getValue());
   }
 
 
@@ -71,27 +56,41 @@ public class PlusDMIndicatorTest extends AbstractIndicatorTest<Num> {
     final var series = new MockBarSeriesBuilder().withNumFactory(this.numFactory).build();
     final var yesterdayBar = series.barBuilder().openPrice(0).closePrice(0).highPrice(6).lowPrice(12).build();
     final var todayBar = series.barBuilder().openPrice(0).closePrice(0).highPrice(12).lowPrice(6).build();
-    final var dup = prepareIndicator(series, yesterdayBar, todayBar);
-    assertNumEquals(0, dup.getValue());
+    final var down = prepareIndicator(series, yesterdayBar, todayBar);
+    assertNumEquals(0, down.getValue());
   }
 
 
   @Test
   public void zeroDirectionalMovement3() {
     final var series = new MockBarSeriesBuilder().withNumFactory(this.numFactory).build();
-    final var yesterdayBar = series.barBuilder().openPrice(0).closePrice(0).highPrice(6).lowPrice(20).build();
+    final var yesterdayBar = series.barBuilder().openPrice(0).closePrice(0).highPrice(6).lowPrice(6).build();
     final var todayBar = series.barBuilder().openPrice(0).closePrice(0).highPrice(12).lowPrice(4).build();
-    final var dup = prepareIndicator(series, yesterdayBar, todayBar);
-    assertNumEquals(0, dup.getValue());
+    final var down = prepareIndicator(series, yesterdayBar, todayBar);
+    assertNumEquals(0, down.getValue());
+  }
+
+
+  private static MinusDMIndicator prepareIndicator(
+      final BacktestBarSeries series, final BacktestBar yesterdayBar,
+      final BacktestBar todayBar
+  ) {
+    final var down = new MinusDMIndicator(series);
+    series.replaceStrategy(new MockStrategy(down));
+    series.addBar(yesterdayBar);
+    series.addBar(todayBar);
+    series.advance();
+    series.advance();
+    return down;
   }
 
 
   @Test
   public void positiveDirectionalMovement() {
     final var series = new MockBarSeriesBuilder().withNumFactory(this.numFactory).build();
-    final var yesterdayBar = series.barBuilder().openPrice(0).closePrice(0).highPrice(6).lowPrice(6).build();
+    final var yesterdayBar = series.barBuilder().openPrice(0).closePrice(0).highPrice(6).lowPrice(20).build();
     final var todayBar = series.barBuilder().openPrice(0).closePrice(0).highPrice(12).lowPrice(4).build();
-    final var dup = prepareIndicator(series, yesterdayBar, todayBar);
-    assertNumEquals(6, dup.getValue());
+    final var down = prepareIndicator(series, yesterdayBar, todayBar);
+    assertNumEquals(16, down.getValue());
   }
 }
