+++ ta4j-core/src/test/resources/org/ta4j/core/indicators/numeric/adx/ADX.xls
